+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 Unreleased
 ***********
+* Adds a `/calibrators` endpoint that returns a list of approved LOW Calibrator targets
 
 9.0.0
 ***********